+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+# 1.10.12
+- Add a new async method share_value_with_data to the BaseTask class
+
 # 1.10.11
 - add support type bytes or bytearray for share memory in BaseTask